Enhance unit tests, support for variables from all stack frames, fix build problems with the scripting plug-in, add support for the standard tcf registers service to the Windows debugger.

+++ edc/org.eclipse.cdt.debug.edc/src/org/eclipse/cdt/debug/edc/internal/StreamBufferBase.java
@@ -83,10 +83,10 @@ public abstract class StreamBufferBase implements IStreamBuffer {
 	/* (non-Javadoc)
 	 * @see org.eclipse.cdt.debug.edc.internal.IStreamBuffer#wrapSubsection(int)
 	 */
-	public IStreamBuffer wrapSubsection(int size) {
+	public IStreamBuffer wrapSubsection(long size) {
 		long availableSize = capacity() - position();
 		if (availableSize < size)
-			size = (int) availableSize;
+			size = availableSize;
 		return createSubBuffer(position() + baseOffset, size);
 	}
